Fix type mismatch in analyze_portfolio_copy_paths function
Cast ROW_NUMBER() results to INT to match function return type.
ROW_NUMBER() returns BIGINT by default, causing "structure of query
does not match function result type" error.

Applied cast to all three CTEs: first_portfolios, last_portfolios, full_sequences

+++ supabase/migrations/20251203_fix_portfolio_copy_paths_dedup.sql
@@ -50,7 +50,7 @@ BEGIN
   first_portfolios AS (
     SELECT
       'first_portfolio'::TEXT as analysis_type,
-      ROW_NUMBER() OVER (ORDER BY COUNT(DISTINCT user_id) DESC) as path_rank,
+      ROW_NUMBER() OVER (ORDER BY COUNT(DISTINCT user_id) DESC)::INT as path_rank,
       ARRAY[portfolio_ticker] as portfolio_sequence,
       COUNT(DISTINCT user_id)::INT as converter_count,
       ROUND((COUNT(DISTINCT user_id)::NUMERIC / total_converters * 100), 2) as pct_of_converters,
@@ -66,7 +66,7 @@ BEGIN
   last_portfolios AS (
     SELECT
       'last_portfolio'::TEXT as analysis_type,
-      ROW_NUMBER() OVER (ORDER BY COUNT(DISTINCT user_id) DESC) as path_rank,
+      ROW_NUMBER() OVER (ORDER BY COUNT(DISTINCT user_id) DESC)::INT as path_rank,
       ARRAY[portfolio_ticker] as portfolio_sequence,
       COUNT(DISTINCT user_id)::INT as converter_count,
       ROUND((COUNT(DISTINCT user_id)::NUMERIC / total_converters * 100), 2) as pct_of_converters,
@@ -109,7 +109,7 @@ BEGIN
   full_sequences AS (
     SELECT
       'full_sequence'::TEXT as analysis_type,
-      ROW_NUMBER() OVER (ORDER BY COUNT(*) DESC) as path_rank,
+      ROW_NUMBER() OVER (ORDER BY COUNT(*) DESC)::INT as path_rank,
       us.portfolio_sequence,
       COUNT(*)::INT as converter_count,
       ROUND((COUNT(*)::NUMERIC / total_converters * 100), 2) as pct_of_converters,
